+++ openpathsampling/storage/storage.py
@@ -8,6 +8,7 @@ import logging
 import time
 
 import openpathsampling as paths
+import openpathsampling.storage.stores.wrapper
 from openpathsampling.netcdfplus import NetCDFPlus, WeakLRUCache, ObjectStore, \
     ImmutableDictStore, NamedObjectStore
 import openpathsampling.engines as peng
@@ -73,7 +74,7 @@ class Storage(NetCDFPlus):
         self.create_store('topologies', NamedObjectStore(peng.Topology))
         self.create_store('cvs', paths.storage.CVStore())
 
-        self.create_store('snapshots', paths.storage.SnapshotWrapperStore())
+        self.create_store('snapshots', openpathsampling.storage.stores.wrapper.SnapshotWrapperStore())
 
         self.create_store('samples', paths.storage.SampleStore())
         self.create_store('samplesets', paths.storage.SampleSetStore())
