+++ openpathsampling/storage/stores/__init__.py
@@ -3,7 +3,8 @@ from mcstep import MCStepStore
 from movechange import MoveChangeStore
 from sample import SampleSetStore, SampleStore
 from snapshot import (
-    SnapshotWrapperStore, SnapshotValueStore, BaseSnapshotStore,
+    SnapshotValueStore, BaseSnapshotStore,
     FeatureSnapshotStore)
+from openpathsampling.storage import SnapshotWrapperStore
 from trajectory import TrajectoryStore
 from pathsimulator import PathSimulatorStore
