+++ openpathsampling/storage/stores/snapshot_wrapper.py
@@ -1,14 +1,12 @@
-import abc
 import logging
 from uuid import UUID
 
 import openpathsampling.engines as peng
-from openpathsampling.netcdfplus import LoaderProxy
-from openpathsampling.netcdfplus import NetCDFPlus, ObjectStore, \
-    LRUChunkLoadingCache
+from openpathsampling.netcdfplus import ObjectStore, with_timing_logging, \
+    NetCDFPlus, LoaderProxy
 
-from openpathsampling.netcdfplus import with_timing_logging
-from openpathsampling.netcdfplus import IndexedObjectStore
+from snapshot_feature import FeatureSnapshotStore
+from snapshot_value import SnapshotValueStore
 
 logger = logging.getLogger(__name__)
 init_log = logging.getLogger('openpathsampling.initialization')
@@ -64,316 +62,6 @@ class ReversalHashedList(dict):
         k = key & ~1
         if k in self:
             dict.__delitem__(self, k)
-
-
-# ==============================================================================
-# ABSTRACT BASE CLASS FOR SNAPSHOTS
-# ==============================================================================
-
-class BaseSnapshotStore(IndexedObjectStore):
-    """
-    An ObjectStore for Snapshots in netCDF files.
-    """
-
-    __metaclass__ = abc.ABCMeta
-
-    def __init__(self, descriptor):
-        """
-
-        Attributes
-        ----------
-        descriptor : openpathsampling.engines.SnapshotDescriptor
-            a descriptor knowing the snapshot class and a dictionary of
-            dimensions and their lengths.
-
-        """
-
-        # Using a store with None as type will not interfere with the main
-        # `SnapshotWrapperStore`
-        super(BaseSnapshotStore, self).__init__(None, json=False)
-        self.descriptor = descriptor
-        self._dimensions = descriptor.dimensions
-        self._cls = self.descriptor.snapshot_class
-
-    @property
-    def snapshot_class(self):
-        return self._cls
-
-    @property
-    def dimensions(self):
-        return self.descriptor['dimensions']
-
-    def __repr__(self):
-        return "store.%s[%s(%s)]" % (
-            self.prefix, self._cls.__name__, 'BaseSnapshot')
-
-    @staticmethod
-    def paired_idx(idx):
-        """
-        Return the paired index
-
-        Snapshots are stored in pairs (2n, 2n+1) where one is the reversed copy.
-        This make storing CVs easier. This function allows to get the paired
-        index or the index of snapshot.reversed
-
-        The implementation uses the trick that all you have to do is flip the
-        lowest bit that determines even or odd.
-
-        Parameters
-        ----------
-        idx : int
-            the one part of the paired index
-
-        Returns
-        -------
-        int
-            the other part of the paired index
-        """
-        return idx ^ 1
-
-    def to_dict(self):
-        return {
-            'descriptor': self.descriptor,
-        }
-
-    def load(self, idx):
-        pos = idx / 2
-
-        # we want to load by uuid and it was not in cache.
-        if pos in self.index:
-            n_idx = self.index[pos]
-        else:
-            raise KeyError(idx)
-
-        if n_idx < 0:
-            return None
-
-        # # if it is in the cache, return it
-        # try:
-        #     obj = self.cache[n_idx]
-        #     if idx & 1:
-        #         obj = obj.reversed
-        #
-        #     return obj
-        #
-        # except KeyError:
-        #     pass
-
-        obj = self._load(n_idx)
-
-        # self.cache[n_idx] = obj
-
-        if idx & 1:
-            obj = obj.reversed
-
-        return obj
-
-    def _load(self, idx):
-        """
-        Load a snapshot from the storage.
-
-        Parameters
-        ----------
-        idx : int
-            the integer index of the snapshot to be loaded
-
-        Returns
-        -------
-        snapshot : :obj:`openpathsampling.engines.BaseSnapshot`
-            the loaded snapshot instance
-        """
-
-        # if not load and return it
-        st_idx = int(idx)
-
-        obj = self._cls.__new__(self._cls)
-        self._cls.init_empty(obj)
-        self._get(st_idx, obj)
-        return obj
-
-    def save(self, obj, idx=None):
-        pos = idx / 2
-
-        if pos in self.index:
-            # has been saved so quit and do nothing
-            return idx
-
-        # n_idx = self.free() / 2
-        n_idx = len(self.index)
-
-        # mark as saved so circular dependencies will not cause infinite loops
-        self.index.append(pos)
-
-        # make sure in nested saving that an IDX is not used twice!
-        # self.reserve_idx(n_idx)
-
-        logger.debug('Saving ' + str(type(obj)) + ' using IDX #' + str(n_idx))
-
-        try:
-            self._save(obj, n_idx)
-            self.vars['index'][n_idx] = pos
-
-            # store the name in the cache
-            # if hasattr(self, 'cache'):
-            self.cache[n_idx] = obj
-
-        except:
-            logger.debug('Problem saving %d !' % n_idx)
-            # in case we did not succeed remove the mark as being saved
-            del self.index[pos]
-            # self.release_idx(n_idx)
-            raise
-
-        # self.release_idx(n_idx)
-        self._set_id(n_idx, obj)
-
-        return idx
-
-    def _save(self, snapshot, idx):
-        """
-        Add the current state of the snapshot in the database.
-
-        Parameters
-        ----------
-        snapshot :class:`openpathsampling.snapshots.AbstractSnapshot`
-            the snapshot to be saved
-        idx : int or None
-            if idx is not None the index will be used for saving in the storage.
-            This might overwrite already existing trajectories!
-
-        Notes
-        -----
-        This also saves all contained frames in the snapshot if not done yet.
-        A single Snapshot object can only be saved once!
-        """
-
-        self._set(idx, snapshot)
-
-    @abc.abstractmethod
-    def _get(self, idx, snapshot):
-        pass
-
-    @abc.abstractmethod
-    def _set(self, idx, snapshot):
-        pass
-
-    # def _get_id(self, idx, obj):
-    #     print idx, obj
-    #     print 1/0
-    #     uuid = self.vars['uuid'][int(idx / 2)]
-    #
-    #     if idx & 1:
-    #         uuid = StorableObject.ruuid(uuid)
-    #
-    #     obj.__uuid__ = uuid
-    #
-    # def _set_id(self, idx, obj):
-    #     self.vars['uuid'][int(idx / 2)] = obj.__uuid__
-
-    def load_indices(self):
-        self.index.extend(self.vars['index'])
-
-    def all(self):
-        return peng.Trajectory(map(self.proxy, range(len(self))))
-
-    # def duplicate(self, snapshot):
-    #     """
-    #     Store a duplicate of the snapshot as new
-    #
-    #     Parameters
-    #     ----------
-    #     snapshot : :class:`openpathsampling.engines.BaseSnapshot`
-    #
-    #     Returns
-    #     -------
-    #     int
-    #         the index used for storing it in the store. This is the
-    #         save as used by save.
-    #
-    #     Notes
-    #     -----
-    #     This will circumvent the caching and indexing completely. This would be
-    #     equivalent of creating a copy of the current snapshot and store this one
-    #     and throw the copy away, leaving the given snapshot untouched. This
-    #     allows you to treat the snapshot as mutual.
-    #
-    #     The use becomes more obvious when applying to storing trajectories.
-    #     The only way to make use of this feature is using the returned `idx`
-    #
-    #     >>> idx = store.duplicate(snap)
-    #     >>> loaded = store[idx]  # return a duplicated as new object
-    #     >>> proxy = paths.LoaderProxy(store, idx) # use duplicate w/o loading
-    #
-    #     """
-    #     idx = len(self.index)
-    #     st_idx = int(idx)
-    #     self._set(st_idx, snapshot)
-    #
-    #     return idx
-
-    def __iter__(self):
-        for idx in range(0, len(self), 2):
-            yield self[idx]
-
-    def __getitem__(self, item):
-        """
-        Enable numpy style selection of object in the store
-        """
-        try:
-            if type(item) is int or type(item) is str or type(item) is UUID:
-                return self.load(item)
-            elif type(item) is slice:
-                return [self.load(idx)
-                        for idx in range(*item.indices(len(self)))]
-            elif type(item) is list:
-                return [self.load(idx) for idx in item]
-            elif item is Ellipsis:
-                return iter(self)
-        except KeyError:
-            return None
-
-    def __len__(self):
-        return len(self.storage.dimensions[self.prefix]) * 2
-
-
-# ==============================================================================
-# FEATURE BASED SINGLE CLASS FOR ALL SNAPSHOT TYPES
-# ==============================================================================
-
-class FeatureSnapshotStore(BaseSnapshotStore):
-    """
-    An ObjectStore for Snapshots in netCDF files.
-    """
-
-    def __init__(self, descriptor):
-        super(FeatureSnapshotStore, self).__init__(descriptor)
-
-    @property
-    def classes(self):
-        return self.snapshot_class.__features__.classes
-
-    @property
-    def storables(self):
-        return self.snapshot_class.__features__.storables
-
-    def _set(self, idx, snapshot):
-        [self.write(attr, idx, snapshot) for attr in self.storables]
-
-    def _get(self, idx, snapshot):
-        [setattr(snapshot, attr, self.vars[attr][idx])
-         for attr in self.storables]
-
-    def initialize(self):
-        super(FeatureSnapshotStore, self).initialize()
-
-        for dim, size in self._dimensions.iteritems():
-            self.storage.create_dimension(self.prefix + dim, size)
-
-        for feature in self.classes:
-            if hasattr(feature, 'netcdfplus_init'):
-                feature.netcdfplus_init(self)
-
-        self.storage.sync()
 
 
 class SnapshotWrapperStore(ObjectStore):
@@ -1076,163 +764,3 @@ class SnapshotWrapperStore(ObjectStore):
             return [self.load(idx) for idx in item]
         elif item is Ellipsis:
             return iter(self)
-
-
-class SnapshotValueStore(ObjectStore):
-    def __init__(
-            self,
-            time_reversible=True,
-            allow_incomplete=False,
-            chunksize=256
-    ):
-        super(SnapshotValueStore, self).__init__(None)
-        self.snapshot_index = None
-        if not time_reversible and not allow_incomplete:
-            raise RuntimeError(
-                'Only time_reversible CVs can currently be '
-                'stored using mode "complete"')
-
-        self.time_reversible = time_reversible
-        self.allow_incomplete = allow_incomplete
-        self.chunksize = chunksize
-
-        self.snapshot_pos = None
-        self._len = 0
-
-    def to_dict(self):
-        return {
-            'time_reversible': self.time_reversible,
-            'allow_incomplete': self.allow_incomplete
-        }
-
-    def create_uuid_index(self):
-        return dict()
-
-    def register(self, storage, prefix):
-        super(SnapshotValueStore, self).register(storage, prefix)
-        # print self.storage.__dict__.keys()
-        self.snapshot_pos = self.storage.stores['snapshots'].pos
-
-    def __len__(self):
-        return len(self.variables['value'])
-
-    # ==========================================================================
-    # LOAD/SAVE DECORATORS FOR CACHE HANDLING
-    # ==========================================================================
-
-    def load(self, idx):
-        pos = self.snapshot_pos(idx)
-        # print idx.__uuid__ in self.storage.stores['snapshots'].index
-        # print self.storage.stores['snapshots'].index[idx.__uuid__]
-        # print self.snapshot_pos
-        # print self.storage.stores['snapshots'].pos
-        # print 'CV:', idx, pos
-
-        if pos is None:
-            return None
-
-        if self.time_reversible:
-            pos /= 2
-
-        if self.allow_incomplete:
-            # we want to load by uuid and it was not in cache.
-            if pos in self.index:
-                n_idx = self.index[pos]
-            else:
-                return None
-            if n_idx < 0:
-                return None
-        else:
-            if pos < self._len:
-                n_idx = pos
-            else:
-                return None
-
-        # if it is in the cache, return it
-        try:
-            obj = self.cache[n_idx]
-            return obj
-
-        except KeyError:
-            pass
-
-        obj = self.vars['value'][n_idx]
-
-        self.cache[n_idx] = obj
-
-        return obj
-
-    def __setitem__(self, idx, value):
-        pos = self.snapshot_pos(idx)
-
-        if pos is None:
-            return
-
-        if self.time_reversible:
-            pos /= 2
-
-        if self.allow_incomplete:
-            if pos in self.index:
-                return
-
-            n_idx = len(self.index)
-
-            self.cache.update_size(n_idx)
-
-        else:
-            if pos < self._len:
-                return
-
-            n_idx = idx
-
-        if self.allow_incomplete:
-            # only if partial storage is used store index and update
-            self.vars['index'][n_idx] = pos
-            self.index[pos] = n_idx
-
-        self.vars['value'][n_idx] = value
-        self.cache[n_idx] = value
-        self._len = max(self._len, n_idx + 1)
-
-    def fill_cache(self):
-        self.cache.load_max()
-
-    def restore(self):
-        if self.allow_incomplete:  # only if partial storage is used
-            for pos, idx in enumerate(self.vars['index'][:]):
-                self.index[idx] = pos
-
-        self._len = len(self)
-        self.initialize_cache()
-
-    def initialize(self):
-        self.initialize_cache()
-
-    def initialize_cache(self):
-        self.cache = LRUChunkLoadingCache(
-            chunksize=self.chunksize,
-            max_chunks=1000,
-            variable=self.vars['value']
-        )
-        self.cache.update_size()
-
-    def __getitem__(self, item):
-        # enable numpy style selection of objects in the store
-        try:
-            if isinstance(item, peng.BaseSnapshot):
-                return self.load(item)
-            elif type(item) is list:
-                return [self.load(idx) for idx in item]
-            elif item is Ellipsis:
-                return iter(self)
-        except KeyError:
-            return None
-
-    def get(self, item):
-        if self.allow_incomplete:
-            try:
-                return self[item]
-            except KeyError:
-                return None
-        else:
-            return self[item]
